requirements.txt
anyio==4.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
boto3==1.34.122
botocore==1.34.131
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
cryptography==42.0.5
defusedxml==0.8.0rc2
deprecation==2.1.0
dj-database-url==2.1.0
Django==5.1.2
django-ckeditor==6.7.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-htmlmin==0.11.0
django-js-asset==3.0.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
django-storages==1.14.3
django-templated-mail==1.1.1
django-tenant-schemas==1.12.0
django-tenants==3.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-yasg==1.21.7
flet==0.19.0
flet-core==0.19.0
flet-django==0.4.5
flet-runtime==0.19.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==0.17.3
httpx==0.24.1
idna==3.7
inflection==0.5.1
Jinja2==3.1.3
jmespath==1.0.1
josepy==1.14.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mozilla-django-oidc==4.0.1
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.0
Pillow==10.1.0
psycopg==3.1.18
psycopg2-binary==2.9.10
pycparser==2.22
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-environ==0.4.54
python-keycloak==4.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.1
repath==0.9.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
